Rewrote receive and send thread logic to use double buffer rings and buffer pools.

+++ src/main/scala/edu/berkeley/velox/net/NetworkService.scala
@@ -2,15 +2,15 @@ package edu.berkeley.velox.net
 
 import edu.berkeley.velox._
 import edu.berkeley.velox.rpc.{MessageService, KryoMessageService}
-import java.util.concurrent.atomic.AtomicInteger
+import java.util.concurrent.atomic.{AtomicLong, AtomicInteger}
 
 trait NetworkService {
   var msgSentCounter = new AtomicInteger(0)
   var msgRecvCounter = new AtomicInteger(0)
-  var bytesWrittenCounter = new AtomicInteger(0)
-  var bytesSentCounter = new AtomicInteger(0)
-  var bytesRecvCounter = new AtomicInteger(0)
-  var bytesReadCounter = new AtomicInteger(0)
+  var bytesWrittenCounter = new AtomicLong(0)
+  var bytesSentCounter = new AtomicLong(0)
+  var bytesRecvCounter = new AtomicLong(0)
+  var bytesReadCounter = new AtomicLong(0)
 
 
   var messageService: MessageService = null
